requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
pyarrow>=12.0.0
pyyaml>=6.0
pydantic>=2.0.0

# Database connectors
sqlalchemy>=2.0.0
pymysql>=1.1.0
google-cloud-bigquery>=3.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Logging and monitoring
structlog>=23.0.0
rich>=13.0.0

# Utilities
click>=8.1.0
python-dotenv>=1.0.0
pathlib2>=2.3.7

# Type checking
types-PyYAML>=6.0.0
types-requests>=2.28.0 